Use grpc providers in conformance tests
We have a load of engine logic in the grpc layer for providers. Skipping that
and going directly to the Provider objects doesn't give an accurate test of how
the engine will work from these tests.
